test(user): helper for register and login
